+++ jdaviz/app.py
@@ -298,6 +298,7 @@ class ApplicationState(State):
     # https://github.com/spacetelescope/jdaviz/pull/3761
     # https://github.com/spacetelescope/jdaviz/pull/3777
     # https://github.com/spacetelescope/jdaviz/pull/3799
+    # https://github.com/spacetelescope/jdaviz/pull/3814
     catalogs_in_dc = CallbackProperty(
         False, docstring="Whether to enable developer mode for adding catalogs to data collection.")
     loader_items = ListCallbackProperty(
@@ -758,6 +759,7 @@ class Application(VuetifyTemplate, HubListener):
                 viewer.center_on(sky_cen)
 
     def _link_new_data_by_component_type(self, new_data_label):
+
         new_data = self.data_collection[new_data_label]
 
         if (new_data.meta.get('_importer') in ('ImageImporter', 'CatalogImporter') and
@@ -808,7 +810,7 @@ class Application(VuetifyTemplate, HubListener):
         any components are compatible with already loaded data. If so, link
         them so that they can be displayed on the same profile1D plot.
         """
-        if self.config in CONFIGS_WITH_LOADERS:
+        if self.config in CONFIGS_WITH_LOADERS and self.config not in ['specviz2d', 'deconfigged']:
             # automatic linking based on component physical types handled by importers
             return
         elif not self.auto_link:
@@ -840,7 +842,6 @@ class Application(VuetifyTemplate, HubListener):
 
             dc.add_link(LinkSame(ref_wavelength_component, linked_wavelength_component))
             return
-
         elif (linked_data.meta.get('Plugin', None) == '3D Spectral Extraction' or
                 (linked_data.meta.get('Plugin', None) == ('Gaussian Smooth') and
                  linked_data.ndim < 3 and  # Cube linking requires special logic. See below
@@ -849,6 +850,12 @@ class Application(VuetifyTemplate, HubListener):
             links = [LinkSame(linked_data.components[0], ref_data.components[0]),
                      LinkSame(linked_data.components[1], ref_data.components[1])]
 
+            dc.add_link(links)
+            return
+        elif (ref_data.ndim == 2 and linked_data.ndim == 1):
+            # Needed for subset linking between 1D and 2D viewers
+            # Spectrum 1D: Pixel Axis 0 [x] <=> Spectrum 2D: Pixel Axis 1 [x]
+            links = [LinkSameWithUnits(linked_data.components[0], ref_data.components[1])]
             dc.add_link(links)
             return
 
@@ -2597,10 +2604,19 @@ class Application(VuetifyTemplate, HubListener):
         if self.config in CONFIGS_WITH_LOADERS:
             data_needing_relinking = []
             for link in self.data_collection.external_links:
-                if data_to_remove == link.data1:
-                    data_needing_relinking.append(link.data2)
-                elif data_to_remove == link.data2:
-                    data_needing_relinking.append(link.data1)
+                if hasattr(link, 'data1') and hasattr(link, 'data2'):
+                    if data_to_remove == link.data1:
+                        data_needing_relinking.append(link.data2)
+                    elif data_to_remove == link.data2:
+                        data_needing_relinking.append(link.data1)
+
+                # ComponentsLink has comp_from and comp_to instead of data1 and data2,
+                # and is currently used in linking for Source Catalog
+                elif hasattr(link, 'comp_from') and hasattr(link, 'comp_to'):
+                    if data_to_remove == link.comp_from.data:
+                        data_needing_relinking.append(link.comp_to.data)
+                    elif data_to_remove == link.comp_to.data:
+                        data_needing_relinking.append(link.comp_from.data)
 
         self.data_collection.remove(data_to_remove)