+++ jdaviz/configs/specviz2d/plugins/spectral_extraction/tests/test_spectral_extraction.py
@@ -521,7 +521,7 @@ class TestTwo2dSpectra:
         # Another 2D Spectrum <=> 2D Spectrum [spectral flux density]
         # Another 2D Spectral Extraction <=> 2D Spectrum [spectral axis]
         # Another 2D Spectral Extraction <=> 2D Spectrum
-        assert len(dc.external_links) == 6
+        assert len(dc.external_links) == 9
         for link in dc.external_links:
             # Check that linking is correct by confirming that both
             # are in `expected_labels`
@@ -562,8 +562,8 @@ class TestTwo2dSpectra:
         min_value_subset = np.min(subset_highlighted_region1)
         max_value_subset = np.max(subset_highlighted_region1)
 
-        expected_min = 0
-        expected_max = 1
+        expected_min = 2
+        expected_max = 7
         atol = 1e-2
         assert_allclose(min_value_subset, expected_min, atol=atol)
         assert_allclose(max_value_subset, expected_max, atol=atol)