+++ CHANGES.rst
@@ -4,6 +4,8 @@
 New Features
 ------------
 
+- Add check for valid model fitting equation operators. [#3818]
+
 - Plugin-specific viewer tools (footprint, catalogs, line lists, line analysis) now have a
   toggle button in the plugin UI to enable/disable a custom toolbar specific to that context. [#3812]
 
@@ -87,6 +89,7 @@ Specviz
 
 Specviz2d
 ^^^^^^^^^
+- Fix subset linking between pixel/wavelength in Specviz2d viewers. [#3817]
 
 4.4.1 (2025-09-26)
 ==================